refactor(modules): rename TokenizerModule -> TokenizerTrainer, remove non-training proxy methods
