+++ formerbox/modules/__init__.py
@@ -1,3 +1,3 @@
-from formerbox.modules.tokenizer_module import TokenizerModule
+from formerbox.modules.tokenizer_trainer import TokenizerTrainer
 from formerbox.modules.transformer_datamodule import TransformerDataModule
 from formerbox.modules.transformer_module import TransformerModule
